First working version of Presentation API shim and HTML Slidy remote
The demo takes the URL of a slide show made with HTML Slidy as input, opens
that slide show on a second screen and controls that slide show afterwards
from the initial screen.

The demo supports Google Cast devices provided it runs under Google Chrome
with the Google Cast extension, and attached second screens provided it
runs using the modified version of Chromium prepared by Intel's Open Source
Technology Center.

A couple of parameters still need to be adjusted (the address of
receiver applications in particular).