Presentation shim: fixed possibility to miss sender message in window session
Previous commit improved the initial exchange of messages between sender and
receiver in the case a presentation is sent to a second browser window. There
was still a possibility that the shim running on the receiver end would not
bind to the "load" event of the window in time (the code checked whether it
ran in a Google Cast session first). This meant the code could never send
the "ispresentation" event in practice.

The shim now waits for the "load" event before it tries to detect whether it
runs in a receiver app or not.